Switched setting wallpaper from qtile to xinitc
Originally, the wallpaper was set using a built in function from one of
the qtile libraries. However, on startup, this usually meant there was a
visible delay between qtile starting, and the wallpaper showing up.
Therefore, switching the wallpaper initialization to the .xinitrc script
allows for the wallpaper to already be loaded when the user loads into
the qtile window manager.